Adminstrative Update README.md
Just added a bit more words to clarify the timer for the administrative, and took out the radio buttons for the administrative part as I feel like that should be on the quiz page side. 
Also took out refresh for the quiz page, as I don't think the page necessarily needs to refresh?